much error fix, much merge branch skript library into master.

+++ src/main/java/io/github/syst3ms/skriptparser/util/DurationUtils.java
@@ -81,7 +81,7 @@ public class DurationUtils {
             if (millis == -1)
                 return Optional.empty();
 
-            duration += delta * millis;
+            duration += (long) (delta * millis);
         }
         return Optional.of(Duration.ofMillis(duration));
     }
